refactor(interview): de-duplicate derivePreselection; drop dead exports
Extract the shared derivePreselection helper into its own module, removing
the identical copy from both AddSiblingWizard and DefineParentsWizard.
Remove the unused getNodeSubjectPossessive export and its test. Un-export
descendantIds, which is only used within parentCandidates.ts.
